nginx-0.1.8-RELEASE import
    *) Bugfix: in the ngx_http_autoindex_module if the long file names were
       in the listing.

    *) Feature: the "^~" modifier in the location directive.

    *) Feature: the proxy_max_temp_file_size directive.

+++ src/http/modules/ngx_http_range_filter.c
@@ -340,7 +340,7 @@ static ngx_int_t ngx_http_range_header_filter(ngx_http_request_t *r)
         return NGX_ERROR;
     }
 
-    boundary = ngx_next_temp_number(0);
+    boundary = (uint32_t) ngx_next_temp_number(0);
 
     /*
      * The boundary header of the range:
